chore(i18n): .js -> .ts
