--- i18n/errorCodes.js
@@ -1,18 +1,0 @@
-export const ERROR_CODES = {
-    ROOM_NOT_FOUND: "errors.room_not_found",
-    PARSE_ERROR: "errors.parse_error",
-    GAME_CLOSED: "errors.game_closed",
-    HOST_QUIT: "errors.host_quit",
-    IMAGE_TOO_LARGE: "errors.image_too_large",
-    CSV_TOO_LARGE: "errors.csv_too_large",
-    CSV_INVALID_FORMAT: "errors.csv_invalid_format",
-    UNKNOWN_FILE_TYPE: "errors.unknown_file_type",
-    SERVER_ERROR: "errors.server_error",
-    BUZZER_REGISTER_ERROR: "errors.buzzer_register",
-    SPECTATOR_REGISTER_ERROR: "errors.spectator_register",
-    CHANGE_LANG_ERROR: "errors.change_lang",
-    CONNECTION_LOST: "errors.connection_lost",
-    MISSING_INPUT: "errors.missing_input",
-    UNABLE_TO_CONNECT: "errors.unable_to_connect",
-    PLAYER_NOT_FOUND: "errors.player_not_found",
-};
