+++ i18n/i18next-scanner.config.ts
